2-animated-image-gif/pyproject.toml
[tool.poetry]
name = "2-animated-image-gif"
version = "0.1.0"
description = ""
authors = ["elskow <elskow@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^10.2.0"
numpy = "^1.26.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
